got the footer structure finished
