+++ python-project-starter/react-app/public/index.html
@@ -1,10 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Simple React App</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <script src="https://kit.fontawesome.com/2ed6a75d45.js" crossorigin="anonymous"></script>
+  <title>Petsy</title>
+</head>
+
+<body>
+  <div id="root"></div>
+</body>
+
 </html>
